feat(eslint-plugin): report only head of function
